apigestionusuariov5/apiredisgest/requirements.txt
# Flask
# redis
# boto3
# pika==1.2.0
# SQLAlchemy
# Flask-SQLAlchemy
# pymysql
# python-dotenv==0.19.1
# Flask-Migrate
# prometheus-flask-exporter==0.18.1
# Werkzeug==1.0.1

Flask==2.0.1
redis
boto3
pika==1.2.0
SQLAlchemy==1.3.0
Flask-SQLAlchemy
pymysql
python-dotenv==0.19.1
Flask-Migrate
prometheus-flask-exporter
MarkupSafe>=2.0.0
Werkzeug==2.0.0
flask-swagger-ui==3.36.0
